Removed validation of presence of content on content-bearing models Closes #28
Fix issues where Travis CI fails because temp dirs and/or files do not exist

+++ spec/spec_helper.rb
@@ -105,8 +105,12 @@ RSpec.configure do |config|
   end
 
   config.after(:suite) do
-    FileUtils.remove_entry_secure(Ddr::Models.external_file_store)
-    FileUtils.remove_entry_secure(Ddr::Models.minter_statefile)
+    if Ddr::Models.external_file_store && Dir.exist?(Ddr::Models.external_file_store)
+      FileUtils.remove_entry_secure(Ddr::Models.external_file_store) 
+    end
+    if Ddr::Models.minter_statefile && File.exist?(Ddr::Models.minter_statefile)
+      FileUtils.remove_entry_secure(Ddr::Models.minter_statefile)
+    end
   end
 
   config.after(:each) do
